Fix multiple interface problem in schema; use module instead
