+++ app/graph/types/dynamic_type.rb
@@ -1,8 +1,6 @@
 class DynamicType < BaseObject
-  implements AnnotationInterface
+  include Types::Inclusions::AnnotationBehaviors
 
-  # TODO: In future version of GraphQL ruby, we can move
-  # this to definition_methods in the annotation interface
   def id
     object.relay_id('dynamic')
   end
